fix: fixed the circular dependency
Also moved helper to service
